Cleaning Warnings
[Change] VulkanEngine.vcxproj, NODEFAULTLIB:library added and warning level set to 4 from 3.

[Change] GUITypes.cpp, FSliders now return false on end.
[Change] CameraController.cpp, casting the m_Camera.Size() to uint32_t.
[Change] DescriptorHandler.cpp, casting the vector.size to uint32_t's.
[Change] DescriptorHandler.h, p_Cache/p_Alloc now nullptr's.
[Change] ObjectManager.h/RenderFactory.h/cpp, m_CurrentRenderID is now a uint32.
[Change] Renderer.h/cpp, UpdateUniformBuffer function doesn't ask for deltatime anymore.
[Change] ShaderManager.cpp, casting vector.size to uint32_t.
[Change] VulkanInitializers.h, CreateDescriptorBufferInfo function, now using a_Offset.
[Change] VulkanDevice.h/cpp/Renderer.cpp, VulkanDeviceSetup function, doesn't require the variable a_Window anymore.
[Change] VulkanDevice.h/cpp/Renderer.cpp, CreateUniformBuffers, doesn't require the variable a_SwapChain anymore

+++ VulkanRenderer/VulkanRendererWindows/MainRenderer/CameraController.cpp
@@ -18,7 +18,7 @@ CameraController::~CameraController()
 void CameraController::AddCamera(glm::vec3 a_Position, glm::vec3 a_Rotationfloat, float a_AspectRatio, float a_NearField, float a_FarField)
 {
 	Transform* cameraTrans = new Transform(a_Position, a_Rotationfloat, glm::vec3(1));
-	CameraObject* camera = new CameraObject(0, cameraTrans, m_Cameras.size(), m_StandardFov, a_AspectRatio, a_NearField, a_FarField);
+	CameraObject* camera = new CameraObject(0, cameraTrans, static_cast<uint32_t>(m_Cameras.size()), m_StandardFov, a_AspectRatio, a_NearField, a_FarField);
 
 	m_Cameras.push_back(camera);
 }
